Move auth stuff into shared auth module
