+++ tsconfig.json
@@ -25,6 +25,7 @@
       "@assets/*": ["./assets/*"],
       "@graphql/*": ["./src/graphql/*"],
       "@lib/*": ["./lib/*"],
+      "@auth/*": ["./auth/*"],
       "env": ["./.env"]
     }
   },
